+++ api-builder-plugin-fn-elasticsearch/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- By default using nodes instead of node to configure an Elasticsearch nodes cluster
 
 ## [1.0.6] 2020-09-10
 ### Changed